api/requirements.txt
sqlalchemy
alembic
flask
python-dotenv
mysql-connector-python
